@@ -1,24 +1,39 @@
-/* eslint-disable no-unused-vars */
-import App from './App.vue'
-import router from './router'
-import './filters'
-import './services/confetti'
-require('./bootstrap')
+/**
+ * First we will load all of this project's JavaScript dependencies which
+ * includes Vue and other libraries. It is a great starting point when
+ * building robust, powerful web applications using Vue and Laravel.
+ */
 
-window.Vue = require('vue')
-// eslint-disable-next-line no-undef
-Vue.prototype.$axios = axios
-// eslint-disable-next-line no-undef
-Vue.prototype.$Echo = Echo
+import "./bootstrap";
+import { createApp } from "vue";
+import router from "./router";
+import App from "./App.vue";
 
-// eslint-disable-next-line no-undef
-const app = new Vue({
-  render: h => h(App),
-  router,
-  data: {
-    user: window.__app__.user, // we always have user if they login as we return user object in app.blade.php
-    rooms: window.__app__.rooms,
-    emojis: window.__app__.emojis,
-    appName: 'Realtime Chat | Laravel, VueJS, Redis, Laravel Echo, SocketIO'
-  }
-}).$mount('#app')
+/**
+ * Next, we will create a fresh Vue application instance. You may then begin
+ * registering components with the application instance so they are ready
+ * to use in your application's views. An example is included for you.
+ */
+
+const app = createApp(App);
+app.use(router);
+
+/**
+ * The following block of code may be used to automatically register your
+ * Vue components. It will recursively scan this directory for the Vue
+ * components and automatically register them with their "basename".
+ *
+ * Eg. ./components/ExampleComponent.vue -> <example-component></example-component>
+ */
+
+// Object.entries(import.meta.glob('./**/*.vue', { eager: true })).forEach(([path, definition]) => {
+//     app.component(path.split('/').pop().replace(/\.\w+$/, ''), definition.default);
+// });
+
+/**
+ * Finally, we will attach the application instance to a HTML element with
+ * an "id" attribute of "app". This element is included with the "auth"
+ * scaffolding. Otherwise, you will need to add an element yourself.
+ */
+
+app.mount("#app");
