@@ -1,14 +1,4 @@
-window._ = require('lodash')
-
-/**
- * We'll load jQuery and the Bootstrap jQuery plugin which provides support
- * for JavaScript based Bootstrap features such as modals and tabs. This
- * code may be modified to fit the specific needs of your application.
- */
-
-try {
-  window.$ = window.jQuery = require('jquery')
-} catch (e) {}
+// import 'bootstrap';
 
 /**
  * We'll load the axios HTTP library which allows us to easily issue requests
@@ -16,9 +6,12 @@ try {
  * CSRF token as a header based on the value of the "XSRF" token cookie.
  */
 
-window.axios = require('axios')
+import axios from 'axios';
+import $ from 'jquery'
+window.axios = axios; // without this line toOthers doesn't work (not really know why??)
+window.$ = $;
 
-window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest'
+axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
 
 /**
  * Echo exposes an expressive API for subscribing to channels and listening
@@ -26,12 +19,28 @@ window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest'
  * allows your team to easily build robust real-time web applications.
  */
 
-// eslint-disable-next-line import/first
+// import Echo from 'laravel-echo';
+
+// import Pusher from 'pusher-js';
+// window.Pusher = Pusher;
+
+// window.Echo = new Echo({
+//     broadcaster: 'pusher',
+//     key: import.meta.env.VITE_PUSHER_APP_KEY,
+//     cluster: import.meta.env.VITE_PUSHER_APP_CLUSTER ?? 'mt1',
+//     wsHost: import.meta.env.VITE_PUSHER_HOST ?? `ws-${import.meta.env.VITE_PUSHER_APP_CLUSTER}.pusher.com`,
+//     wsPort: import.meta.env.VITE_PUSHER_PORT ?? 80,
+//     wssPort: import.meta.env.VITE_PUSHER_PORT ?? 443,
+//     forceTLS: (import.meta.env.VITE_PUSHER_SCHEME ?? 'https') === 'https',
+//     enabledTransports: ['ws', 'wss'],
+// });
+
 import Echo from 'laravel-echo'
 
-window.io = require('socket.io-client')
+// window.io comes from laravel-echo-server, so that we have compatible version
+// window.io = require('socket.io-client')
 
 window.Echo = new Echo({
   broadcaster: 'socket.io',
-  host: `${window.location.protocol}//${window.location.hostname}:${process.env.MIX_FRONTEND_PORT}`
+  host: `${window.location.protocol}//${window.location.hostname}:${import.meta.env.VITE_LARAVEL_ECHO_SERVER_PORT}`
 })
