@@ -9,7 +9,7 @@ use Illuminate\Broadcasting\PrivateChannel;
 use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Queue\SerializesModels;
-use App\Message;
+use App\Models\Message;
 
 class MessagePosted implements ShouldBroadcast
 {
@@ -29,14 +29,17 @@ class MessagePosted implements ShouldBroadcast
     /**
      * Get the channels the event should broadcast on.
      *
-     * @return \Illuminate\Broadcasting\Channel|array
+     * @return array<int, \Illuminate\Broadcasting\Channel>
      */
     public function broadcastOn()
     {
         if (strpos($this->message->room, '__') !== false) { // must use !== false
-            return new PrivateChannel('room.'.$this->message->receiver);
+            return [
+                new PrivateChannel('room.'.$this->message->receiver),
+                new PrivateChannel('room.'.$this->message->room)
+            ];
         } else {
-            return new PresenceChannel('room.'.$this->message->room);
+            return [new PresenceChannel('room.'.$this->message->room)];
         }
     }
 }
