@@ -9,21 +9,17 @@ use Illuminate\Broadcasting\PrivateChannel;
 use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Queue\SerializesModels;
-use Illuminate\Contracts\Broadcasting\ShouldBroadcastNow;
 
 class MessageReacted implements ShouldBroadcast
 {
-    use Dispatchable, InteractsWithSockets;
-
-    /**
-     * Create a new event instance.
-     *
-     * @return void
-     */
+    use Dispatchable, InteractsWithSockets, SerializesModels;
 
     public $reaction;
     public $channel;
 
+    /**
+     * Create a new event instance.
+     */
     public function __construct($reaction, $channel)
     {
         $this->reaction = $reaction;
@@ -33,14 +29,14 @@ class MessageReacted implements ShouldBroadcast
     /**
      * Get the channels the event should broadcast on.
      *
-     * @return \Illuminate\Broadcasting\Channel|array
+     * @return array<int, \Illuminate\Broadcasting\Channel>
      */
     public function broadcastOn()
     {
         if (strpos($this->channel, '__') !== false) { // must use !== false
-            return new PrivateChannel('room.'.$this->channel);
+            return [new PrivateChannel('room.'.$this->channel)];
         } else {
-            return new PresenceChannel('room.'.$this->channel);
+            return [new PresenceChannel('room.'.$this->channel)];
         }
     }
 }
