@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Message;
 use Auth;
+use App\Models\Message;
+use App\Models\Reaction;
 use App\Events\MessagePosted;
+use App\Events\MessageReacted;
 
 class MessageController extends Controller
 {
@@ -35,5 +37,30 @@ class MessageController extends Controller
         broadcast(new MessagePosted($message->load(['sender', 'reactions.user'])))->toOthers(); // send to others EXCEPT user who sent this message
 
         return response()->json(['message' => $message->load(['sender', 'reactions.user'])]);
+    }
+
+    public function react (Request $request) {
+        $reaction = Reaction::where(['msg_id' => $request->input('msg_id'), 'user_id' => $request->input('user_id')])->first();
+
+        if ($reaction) {
+            if ($reaction->emoji_id === $request->input('emoji_id')) {
+                $reaction->delete();
+            } else {
+                $reaction->emoji_id = $request->input('emoji_id');
+
+                $reaction->save();
+            }
+        } else {
+            $reaction = new Reaction();
+            $reaction->msg_id = $request->input('msg_id');
+            $reaction->user_id = $request->input('user_id');
+            $reaction->emoji_id = $request->input('emoji_id');
+
+            $reaction->save();
+        }
+
+        broadcast(new MessageReacted($reaction->load('user'), $reaction->message->room))->toOthers();
+
+        return ['success' => true, 'reaction' => $reaction];
     }
 }
