@@ -1,48 +1,58 @@
+<script setup>
+import { computed, inject } from "vue";
+
+const props = defineProps({
+  reactions: {
+    type: Array,
+    default: []
+  },
+});
+
+const emojis = inject('$emojis')
+
+const reactionFormat = computed(() => {
+  if (!props.reactions.length) {
+    return []
+  }
+  
+  const listReactionEmojis = [];
+
+  for (const e of emojis) {
+    const index = props.reactions.findIndex((r) => r.emoji_id === e.id);
+
+    if (index > -1) {
+      const indexInList = listReactionEmojis.findIndex(
+        (item) => item.id === e.id
+      );
+
+      if (indexInList === -1) {
+        listReactionEmojis.push(e);
+      }
+    }
+  }
+
+  return listReactionEmojis;
+});
+</script>
+
 <template>
-  <div class="reaction-container d-flex justify-content-center align-items-center">
-    <div
-      class="reaction-item d-flex"
-      v-for="e in reactionFormat"
-      :key="e.id"
-    >
-      <img :src="e.src" :alt="e.alt">
+  <div
+    class="reaction-container d-flex justify-content-center align-items-center"
+  >
+    <div class="reaction-item d-flex" v-for="e in reactionFormat" :key="e.id">
+      <img :src="e.src" :alt="e.alt" />
     </div>
     <div class="reaction-item d-flex">
-      <span class="total ml-1">
+      <span class="total ms-1">
         {{ reactions.length }}
       </span>
     </div>
   </div>
 </template>
 
-<script>
-export default {
-  props: ['reactions'],
-  computed: {
-    reactionFormat () {
-      const listReactionEmojis = []
-
-      for (const e of this.$root.emojis) {
-        const index = this.reactions.findIndex(r => r.emoji_id === e.id)
-
-        if (index > -1) {
-          const indexInList = listReactionEmojis.findIndex(item => item.id === e.id)
-
-          if (indexInList === -1) {
-            listReactionEmojis.push(e)
-          }
-        }
-      }
-
-      return listReactionEmojis
-    }
-  }
-}
-</script>
-
 <style lang="scss">
 .reaction-container {
-  box-shadow: 0 2px 4px rgba(0, 0, 0, .15);
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.15);
   background: #fff;
   border-radius: 15px;
   position: absolute;
